Added new file containing several functions used in app, and improved filename+variable readability
